BAU: Cleaned up code
Removed dead code and duplicates
