+++ src/main/java/uk/gov/ida/notification/pki/CredentialBuilder.java
@@ -1,6 +1,5 @@
 package uk.gov.ida.notification.pki;
 
-import java.security.KeyPair;
 import java.security.PrivateKey;
 import java.security.PublicKey;
 
@@ -23,11 +22,7 @@ public class CredentialBuilder {
         );
     }
 
-    public static CredentialBuilder withKeyPair(KeyPair keyPair) {
-        return new CredentialBuilder(keyPair.getPublic(), keyPair.getPrivate());
-    }
-
-    protected CredentialBuilder(PublicKey publicKey, PrivateKey privateKey) {
+    private CredentialBuilder(PublicKey publicKey, PrivateKey privateKey) {
         this.publicKey = publicKey;
         this.privateKey = privateKey;
     }
