+++ src/main/java/uk/gov/ida/notification/saml/converters/EidasSamlParamConverter.java
@@ -1,0 +1,40 @@
+package uk.gov.ida.notification.saml.converters;
+
+import com.google.common.base.Strings;
+import io.dropwizard.jersey.errors.ErrorMessage;
+import org.glassfish.jersey.internal.util.Base64;
+import uk.gov.ida.notification.saml.SamlParser;
+
+import javax.ws.rs.core.Response;
+import javax.ws.rs.ext.ParamConverter;
+
+class EidasSamlParamConverter<T> implements ParamConverter<T> {
+    private SamlParser samlParser;
+
+    EidasSamlParamConverter(SamlParser samlParser) {
+        this.samlParser = samlParser;
+    }
+
+    @Override
+    public T fromString(String value) {
+        if (Strings.isNullOrEmpty(value)) {
+            return null;
+        }
+
+        String decodedSaml = Base64.decodeAsString(value);
+        T parsedSaml = samlParser.parseSamlString(decodedSaml);
+        return parsedSaml;
+    }
+
+    @Override
+    public String toString(T value) {
+        return value.toString();
+    }
+
+    protected Response getErrorResponse(String message) {
+        return Response
+                .status(400)
+                .entity(new ErrorMessage(400, message))
+                .build();
+    }
+}
