+++ src/main/java/uk/gov/ida/notification/saml/converters/ResponseParameterProvider.java
@@ -1,11 +1,7 @@
 package uk.gov.ida.notification.saml.converters;
 
-import com.google.common.base.Strings;
-import io.dropwizard.jersey.errors.ErrorMessage;
-import org.glassfish.jersey.internal.util.Base64;
 import uk.gov.ida.notification.saml.SamlParser;
 
-import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.ParamConverter;
 import javax.ws.rs.ext.ParamConverterProvider;
 import javax.ws.rs.ext.Provider;
@@ -27,30 +23,6 @@ public class ResponseParameterProvider implements ParamConverterProvider {
         if (!org.opensaml.saml.saml2.core.Response.class.equals(rawType)) {
             return null;
         }
-
-        return new ParamConverter<T>() {
-            @Override
-            public T fromString(String value) {
-                if (Strings.isNullOrEmpty(value)) {
-                    return null;
-                }
-
-                String decodedSaml = Base64.decodeAsString(value);
-                T authnRequest = samlParser.parseSamlString(decodedSaml);
-                return authnRequest;
-            }
-
-            @Override
-            public String toString(T value) {
-                return value.toString();
-            }
-
-            protected Response getErrorResponse(String message) {
-                return Response
-                    .status(400)
-                    .entity(new ErrorMessage(400, message))
-                    .build();
-            }
-        };
+        return new EidasSamlParamConverter<>(samlParser);
     }
 }
